+++ tfjs-backend-webgpu/src/kernels/mirror_pad_webgpu.ts
@@ -49,7 +49,7 @@ export class MirrorPadProgram implements WebGPUProgram {
     this.xShape = xShape;
     paddings.map((_, i) => {
       this.uniforms += ` ivec2 pad${i};`;
-      this.uniformsWgsl += ` pad${i} : vec2<u32>;`;
+      this.uniformsWgsl += ` pad${i} : vec2<i32>;`;
     });
     this.offset = mode === 'reflect' ? 0 : 1;
     this.shaderKey = `mirrorPad_${mode}`;
@@ -102,10 +102,11 @@ export class MirrorPadProgram implements WebGPUProgram {
   getUserCodeWgsl(): string {
     const rank = this.xShape.length;
     // The length of paddings are same with the rank of the input tensor.
-    const start = this.xShape.map((_, i) => `uniforms.pad${i}[0]`).join(',');
+    const start =
+        this.xShape.map((_, i) => `u32(uniforms.pad${i}[0])`).join(',');
     const end = this.xShape
                     .map(
-                        (_, i) => `uniforms.pad${i}[0] + uniforms.xShape${
+                        (_, i) => `u32(uniforms.pad${i}[0]) + uniforms.xShape${
                             rank > 1 ? `[${i}]` : ''}`)
                     .join(',');
